refactor: move FAQ component
