+++ packages/site/src/features/Landing/index.tsx
@@ -1,13 +1,11 @@
 import React, { type ReactNode } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import {
-	Accordion,
 	Box,
 	Heading,
 	Image,
 	Separator,
 	SimpleGrid,
-	Span,
 	Stack,
 	VStack,
 } from '@chakra-ui/react';
@@ -20,6 +18,7 @@ import { useLocalePath } from '../../components/Locale';
 import { Text } from '../../components/Text';
 import { TheRock } from '../../components/TheRock';
 
+import { FAQ } from './FAQ';
 import screenshot from './screenshots/app.png';
 import encryptionScreenshot from './screenshots/encryption.png';
 import historyScreenshot from './screenshots/history.png';
@@ -27,129 +26,6 @@ import linksScreenshot from './screenshots/links.png';
 import remindersScreenshot from './screenshots/reminders.png';
 import tagsScreenshot from './screenshots/tags.png';
 import workspacesScreenshot from './screenshots/workspaces.png';
-
-// TODO: move to another file
-export const FAQ = () => {
-	const { t } = useTranslation('faq');
-
-	const faq = [
-		{
-			value: 'betterThanOthers',
-			title: t('betterThanOthers.title'),
-			text: (
-				<Trans
-					t={t}
-					i18nKey="betterThanOthers.text"
-					components={{
-						security: <Link href="/introduction/security/" />,
-					}}
-				/>
-			),
-		},
-		{
-			value: 'differentFromOthers',
-			title: t('differentFromOthers.title'),
-			text: <Trans t={t} i18nKey="differentFromOthers.text" />,
-		},
-		{
-			value: 'whyVaults',
-			title: t('whyVaults.title'),
-			text: <Trans t={t} i18nKey="whyVaults.text" />,
-		},
-		{
-			value: 'workspacesVsVaults',
-			title: t('workspacesVsVaults.title'),
-			text: <Trans t={t} i18nKey="workspacesVsVaults.text" />,
-		},
-		{
-			value: 'workspacesVsFolders',
-			title: t('workspacesVsFolders.title'),
-			text: <Trans t={t} i18nKey="workspacesVsFolders.text" />,
-		},
-		{
-			value: 'stayingOrganized',
-			title: t('stayingOrganized.title'),
-			text: <Trans t={t} i18nKey="stayingOrganized.text" />,
-		},
-		{
-			value: 'ergonomics',
-			title: t('ergonomics.title'),
-			text: <Trans t={t} i18nKey="ergonomics.text" />,
-		},
-		{
-			value: 'replaceMultipleApps',
-			title: t('replaceMultipleApps.title'),
-			text: <Trans t={t} i18nKey="replaceMultipleApps.text" />,
-		},
-		{
-			value: 'nestedTags',
-			title: t('nestedTags.title'),
-			text: <Trans t={t} i18nKey="nestedTags.text" />,
-		},
-		{
-			value: 'connectedInformation',
-			title: t('connectedInformation.title'),
-			text: <Trans t={t} i18nKey="connectedInformation.text" />,
-		},
-		{
-			value: 'privacy',
-			title: t('privacy.title'),
-			text: <Trans t={t} i18nKey="privacy.text" />,
-		},
-		{
-			value: 'openSource',
-			title: t('openSource.title'),
-			text: <Trans t={t} i18nKey="openSource.text" />,
-		},
-		{
-			value: 'personalOrWork',
-			title: t('personalOrWork.title'),
-			text: <Trans t={t} i18nKey="personalOrWork.text" />,
-		},
-		{
-			value: 'largeVaults',
-			title: t('largeVaults.title'),
-			text: <Trans t={t} i18nKey="largeVaults.text" />,
-		},
-		{
-			value: 'oneVault',
-			title: t('oneVault.title'),
-			text: <Trans t={t} i18nKey="oneVault.text" />,
-		},
-		{
-			value: 'markdownFiles',
-			title: t('markdownFiles.title'),
-			text: <Trans t={t} i18nKey="markdownFiles.text" />,
-		},
-	];
-
-	return (
-		<Accordion.Root
-			variant="subtle"
-			size="lg"
-			multiple
-			defaultValue={[
-				'betterThanOthers',
-				'differentFromOthers',
-				'workspacesVsVaults',
-			]}
-		>
-			{faq.map((item, index) => (
-				<Accordion.Item key={index} value={item.value}>
-					<Accordion.ItemTrigger height="4rem">
-						<Span flex="1">{item.title}</Span>
-						<Accordion.ItemIndicator />
-					</Accordion.ItemTrigger>
-					<Accordion.ItemContent>
-						<Accordion.ItemBody fontSize="1.2rem" whiteSpace="pre-line">
-							{item.text}
-						</Accordion.ItemBody>
-					</Accordion.ItemContent>
-				</Accordion.Item>
-			))}
-		</Accordion.Root>
-	);
-};
 
 export default WithLayout(() => {
 	const analytics = useAnalytics();
